Auth services adjustments.
Changed the authentication for a new user on Auth0: instead
of using a post-registration hook, the app now uses a custom
field to display a number of logins (in the JWT).
For that, the following changes are introduced to the app code:
-Added auth requirement for the GET users' endpoint. This
ensures the new JWT is verified before the user gets any
user data.
-Added an authentication header to all users-related
requests.
-Added a function to parse the token payload.
-Added a function to check whether it's the user's
first login. If it is, it's a new user, so the user's
data needs to be added to the database (via POST
request). Since this endpoint requires authentication,
only if the JWT is verified the user's data is added to
the database.
-If it's the user's first login, sends a request for user
data after adding the user to the database. If it's not
the first login, skips the POST request and just gets
the user's data.
-The user's Auth0id is now taken from the JWT payload.
-The authentication variable is now only set after the
JWT is verified by the backend.