Replaced the pagination mechanism
The previous pagination mechanism required fetching all the records from
the database before doing any sort of pagination. This is extremely
inefficient in large numbers and it makes more sense to do
it in the database. For the time being, I've replaced the previous
mechanism with Flask-SQLAlchemy's built in pagination. While it also
has known performance issues, it's still considerably
better than doing it all in memory in Python.